style(workspace): apply clippy's machine-applicable fixes across 25 files (T0626)
With the ~24k detached-doc warnings cleared, `cargo clippy --workspace
--all-targets` surfaced ~400 real findings. This lands the subset clippy can
fix and verify itself (`--fix`), covering the bulk mechanical classes:
collapsible `if`/`else` chains, references immediately dereferenced by the
compiler, needless full-iterator scans where a direct search exists, redundant
same-type casts, manual `Default`/unit-struct construction, redundant pattern
matching, and `items_after_test_module`.

The largest diff (archive_ctx_loader.rs, 2287 lines) is
`items_after_test_module` relocating the test module to the end of the file.
Verified as a pure reordering — identical line multiset AND identical byte
count vs HEAD — and Rust item order is semantically irrelevant.

Gated: cargo test -p verum_vbc --features codegen (minus the slow
stdlib-compile fixtures) = 1878 passed, exactly matching the pre-change
baseline, with only the 2 failures already proven pre-existing (T0627).

+++ crates/llvm/verum_mlir/src/diagnostic/severity.rs
@@ -31,7 +31,7 @@ impl TryFrom<MlirDiagnosticSeverity> for DiagnosticSeverity {
             MlirDiagnosticSeverity_MlirDiagnosticNote => Self::Note,
             MlirDiagnosticSeverity_MlirDiagnosticRemark => Self::Remark,
             MlirDiagnosticSeverity_MlirDiagnosticWarning => Self::Warning,
-            _ => return Err(Error::UnknownDiagnosticSeverity(severity as u32)),
+            _ => return Err(Error::UnknownDiagnosticSeverity(severity)),
         })
     }
 }
